First attempt for simple task-based passes
